Stepの削減とUI調整
